Tweaks from PR review
